Practica_Clases/practicafsm/pedidos/api/cliente_viewset.py
from rest_framework import serializers, viewsets

from pedidos.models import Cliente


class ClienteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cliente
        fields = '__all__'


class ClienteViewSet(viewsets.ModelViewSet):
    serializer_class = ClienteSerializer
    queryset = Cliente.objects.all()
